+++ js/app.config.js
@@ -1,6 +1,6 @@
 app.config = {
-	env						: 'prod',
-	version				:  '1.8.86', // read CHANGELOG
+	env						: 'prod', // 
+	version				:  '1.8.88', // Fix no-scroll in app   !!!!REMEMBER TO CHANGE is_production and is_simulator!!!!!
 	disconnect_timeout_to_stop_polling: 45000, // stop trying to find tables after 45 seconds
 	extended_timeout_to_stop_polling: 90000, // stop trying to find tables when rebooting
 	show_tg				: false, // for testing
@@ -13,13 +13,21 @@ app.config = {
 			port		: 3001,
 		},
 		beta: {		// tests local network
+<<<<<<< HEAD
 			base_url	: 'http://app.dev.withease.io:3001/', //local url
 			api_url		: 'http://dev.webcenter.sisyphus-industries.com/',
 			web_url		: 'http://dev.webcenter.sisyphus-industries.com/',
 			// api_url		: 'https://webcenter.sisyphus-industries.com/',
 			// web_url		: 'https://webcenter.sisyphus-industries.com/',
+=======
+			base_url		: 'http://app.dev.withease.io:3001/', //local url
+			// api_url		: 'https://webcenter.sisyphus-industries.com/', // PRODUCTION
+			// web_url		: 'https://webcenter.sisyphus-industries.com/', // PRODUCTION
+			api_url			: 'http://dev.webcenter.sisyphus-industries.com/', // DEV
+			web_url			: 'http://dev.webcenter.sisyphus-industries.com/', // DEV
+>>>>>>> staging
 			// api_url		: 'http://localhost:3333/',// add entry in your computers /etc/hosts mapped to your bot's IP address
-			// web_url		: 'http://localhost:3333/', //web_center url	***Change to this for Rails web_center= http://localhost:3333/  (aka rails s) //  10.0.0.3	beta_bot.local
+			// web_url			: 'http://localhost:3333/', //web_center url	***Change to this for Rails web_center= http://localhost:3333/  (aka rails s) //  10.0.0.3	beta_bot.local
 			sisbot_url  : 'http://192.168.0.19:3002', //talking to sisbot    //  ... or just put your URL in here '192.168.XX.XXX:3002' << for local Dev Env --insert your ip address + 3000
 			port		: 3001,
 			show_tg		: true